% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sample_study}
\alias{sample_study}
\title{sample_study}
\usage{
sample_study(X, betag, sigb, rho, taug, sigt, OmegaG, lkj_local)
}
\arguments{
\item{X}{covariate data}

\item{betag}{matrix global regression parameters mean}

\item{sigb}{matrix global regression parameters SD}

\item{rho}{interpolant between local and global correlations}

\item{taug}{global tau mean(s)}

\item{sigt}{global tau SD(s)}

\item{OmegaG}{global correlation parameters}

\item{lkj_local}{local correlation LKJ prior parameter}
}
\value{
matrix of responses
}
\description{
Simulate responses for single study
}
\details{
Simulate responses for single study given covariate data.
}
\author{
Pete Dodd
}
