% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muvamere-package.R
\docType{package}
\name{muvamere-package}
\alias{muvamere-package}
\alias{muvamere}
\title{The 'muvamere' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
\author{
\strong{Maintainer}: Pete Dodd \email{p.j.dodd@sheffield.ac.uk} (\href{https://orcid.org/0000-0001-5825-9347}{ORCID})

Authors:
\itemize{
  \item Pete Dodd \email{p.j.dodd@sheffield.ac.uk} (\href{https://orcid.org/0000-0001-5825-9347}{ORCID})
}

}
