% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{mvn_infer}
\alias{mvn_infer}
\title{mvn_infer}
\usage{
mvn_infer(
  Y,
  X,
  Z,
  beta_prior_sd = 5,
  tau_prior_sd = 2.5,
  lkj_prior_scale = 2,
  iter = 2000,
  cores = 4,
  chains = 4,
  ...
)
}
\arguments{
\item{Y}{multivariate response data - each line sign/symptom values for a patient}

\item{X}{multivariable explanatory data - each line a set of predictors for a patient}

\item{Z}{new X to predict/simulate on (TODO)}

\item{beta_prior_sd}{prior for Betas, default=5}

\item{tau_prior_sd}{= 2.5,}

\item{lkj_prior_scale}{LKJ prior scale for correlation, default=2}

\item{iter}{iterations for MCMC, default = 2e3}

\item{cores}{number of cores to use}

\item{chains}{number of chains to use}

\item{prior}{for tau Cauchy scale in correlations, default=2.5}
}
\value{
a Stan sample object
}
\description{
MCMC sampling for data from single study
}
\details{
This generates a Stan sample for MVN data from a single study
}
\author{
Pete Dodd
}
