% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{mvn_infer_mlm}
\alias{mvn_infer_mlm}
\title{mvn_infer_mlm}
\usage{
mvn_infer_mlm(
  Y,
  X,
  study,
  betaM_prior_sd = 1,
  betaS_prior_sd = 0.5,
  tauM_prior_sd = 1,
  tauS_prior_sd = 0.5,
  lkj_local_prior_scale = 3,
  lkj_global_prior_scale = 2,
  rhoA = 2,
  rhoB = 2,
  iter = 2000,
  cores = 4,
  chains = 4,
  ...
)
}
\arguments{
\item{Y}{multivariate response data - each line sign/symptom values for a
patient}

\item{X}{multivariable explanatory data - each line a set of predictors for
a patient}

\item{study}{vector of which study each record belongs to}

\item{betaM_prior_sd}{prior for Betas}

\item{betaS_prior_sd}{prior for Betas}

\item{tauM_prior_sd}{prior for tau}

\item{tauS_prior_sd}{prior for tau}

\item{lkj_local_prior_scale}{prior for local cor}

\item{lkj_global_prior_scale}{prior for global cor}

\item{rhoA}{beta parameter for rho}

\item{rhoB}{beta parameter for rho}

\item{iter}{iterations for MCMC, default}

\item{cores}{number of cores to use}

\item{chains}{number of chains to use}
}
\value{
a Stan sample object
}
\description{
MCMC sampling for data from multiple studies
}
\details{
This generates a Stan sample for MVN data from a multiple studies
}
\author{
Pete Dodd
}
